source/rule_engine/requirements.txt
requests
pika
psycopg2-binary
fastapi
uvicorn
pydantic